스킬 작업중

+++ Assets/02.Scripts/SystemValue.cs
@@ -21,7 +21,7 @@ public static class SystemValue
     public static int RUNE_DAILY_LIMIT = 3;
     public static int RUNE_DURATUIN = 360 * 1;
 
-    public static int SKILL_DISTANCE_BLOCK_SIZE = 1;
+    public static float SKILL_DISTANCE_BLOCK_SIZE = 1f;
     public static int SKILL_MAX_SLOT_COUNT = 6;
     public static int SKILL_MAX_LEVEL = 10;
